Rename some delclarations
