+++ src/pages/Profile/Profile.tsx
@@ -6,7 +6,7 @@ import { useNavigate } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 
 import { db } from "../../context/firebaseSDK";
-import getProjects from "../../utils/getProjects";
+import getProjects from "../../utils/getUserProjects";
 import { AuthContext } from "../../context/authContext";
 
 interface Prop {
