+++ src/utils/getFriendsProjects.tsx
@@ -34,10 +34,10 @@ export default async function getFriendsProjects(
     usersRef,
     where("uid", "in", [userId, ...friendList])
   );
-  const firstFriendquerySnapshot = await getDocs(firstFriendQuery);
-  const firstFriendFetchedProjects: FetchedProjectsType[] = [];
-  firstFriendquerySnapshot.forEach((project) => {
-    firstFriendFetchedProjects.unshift({
+  const friendQuerySnapshot = await getDocs(firstFriendQuery);
+  const fetchedFriendProjects: FetchedProjectsType[] = [];
+  friendQuerySnapshot.forEach((project) => {
+    fetchedFriendProjects.unshift({
       name: "",
       avatar: "",
       projectId: project.id,
@@ -49,16 +49,16 @@ export default async function getFriendsProjects(
     });
   });
   await Promise.all(
-    firstFriendFetchedProjects.map(async (project, index) => {
+    fetchedFriendProjects.map(async (project, index) => {
       const docSnap = await getDoc(doc(db, "users", project.uid));
       const { name, avatar } = docSnap.data() as {
         name: string;
         avatar: string;
       };
-      firstFriendFetchedProjects[index].name = name;
-      firstFriendFetchedProjects[index].avatar = avatar;
+      fetchedFriendProjects[index].name = name;
+      fetchedFriendProjects[index].avatar = avatar;
     })
   );
 
-  return firstFriendFetchedProjects;
+  return fetchedFriendProjects;
 }
